[REFACTOR] css & 도메인 하위 컴포넌트 생성
