+++ .eslintrc
@@ -21,7 +21,6 @@
   "rules": {
     "react-hooks/rules-of-hooks": "error",
     "react-hooks/exhaustive-deps": "warn",
-    "no-undef": "off",
     "react/jsx-uses-react": "off",
     "react/react-in-jsx-scope": "off",
     "@emotion/pkg-renaming": "error",
