+++ jsconfig.json
@@ -1,6 +1,5 @@
 {
   "compilerOptions": {
-    "module": "ES6",
     "jsx": "preserve",
     "paths": {
       "@assets/*": ["./src/assets/*"],
