+++ src/components/My/Intro/index.js
@@ -1,0 +1,6 @@
+import Profile from './Profile';
+import Membership from './MemberShip';
+import Infos from './Infos';
+import SideInfos from './SideInfos';
+
+export { Profile, Membership, Infos, SideInfos };
